requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.7.2
google-cloud-aiplatform==1.52.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3
